requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
bcrypt==5.0.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
idna==3.11
invoke==2.2.1
multidict==6.7.0
paramiko==4.0.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyNaCl==1.6.2
python-dotenv==1.2.1
qiniu==7.17.0
requests==2.32.5
rsa==4.9.1
storage==0.0.4.3
typing_extensions==4.15.0
upload==0.1.1
urllib3==2.6.2
yarl==1.22.0
